Safari-611.1.10/LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-time-control-styles-expected.txt

Testing the MacOSFullscreenMediaControls computed styles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(mediaControls.timeControl.element).width became "448px"
PASS window.getComputedStyle(mediaControls.timeControl.element).left is "10px"
PASS window.getComputedStyle(mediaControls.timeControl.element).top is "49px"
PASS window.getComputedStyle(mediaControls.timeControl.scrubber.element).top is "0px"
PASS window.getComputedStyle(mediaControls.timeControl.elapsedTimeLabel.element).fontSize is "12px"
PASS window.getComputedStyle(mediaControls.timeControl.remainingTimeLabel.element).fontSize is "12px"

PASS successfullyParsed is true

TEST COMPLETE
